fix: Compiler warning fixes and add macOS platform files
- Fix implicit conversion warnings with static_cast throughout codebase
- Add macOS platform implementation files (platform_file_ops.cpp, system_integration.cpp)
- Add macOS Info.plist for app bundle configuration
- Fix size_t to int conversions in archive_handler, file_manager, and theme_manager
- Improve hash_calculator worker thread initialization with proper static_cast

These changes ensure clean compilation on macOS without warnings and complete
the platform-specific implementation layer.

+++ src/core/archive_handler.cpp
@@ -150,8 +150,8 @@ QList<ArchiveFileInfo> ArchiveHandler::scanArchive(const QString& archivePath)
         
         m_totalArchivesProcessed++;
         m_totalFilesScanned += files.size();
-        
-        emit scanCompleted(archivePath, files.size());
+
+        emit scanCompleted(archivePath, static_cast<int>(files.size()));
         qDebug() << "Archive scan completed:" << archivePath << "Files found:" << files.size();
         
     } catch (const std::exception& e) {
@@ -335,7 +335,7 @@ QList<ArchiveFileInfo> ArchiveHandler::scanZipArchive(const QString& archivePath
         }
         
         // Extract filename (everything after the time field)
-        int nameIndex = line.lastIndexOf(parts[3]) + parts[3].length();
+        int nameIndex = static_cast<int>(line.lastIndexOf(parts[3])) + static_cast<int>(parts[3].length());
         if (nameIndex >= line.length()) {
             continue;
         }
@@ -441,7 +441,7 @@ QList<ArchiveFileInfo> ArchiveHandler::scanTarArchive(const QString& archivePath
             continue; // Skip directories
         }
         
-        emit scanProgress(archivePath, ++fileCount, lines.size(), extractedFileName);
+        emit scanProgress(archivePath, ++fileCount, static_cast<int>(lines.size()), extractedFileName);
         
         // Get detailed file info using tar -tvf
         QProcess detailProcess;
